Update Kimbravelator.scpt
Updated main.scpt to get it working again after Kimble updated their app.  Mainly they just renamed all of the field tag IDs, but they added a couple of new ones too.  Note that this version does not now allow you to chose an Excel input file.  I have added the field names to a list in the source now rather than inline as it makes it a little easier to maintain.  Still requires some tidying up, and the re-addition of the file chooser feature.  Although, I am in two minds about that - possibly an unnecessary feature.

Just had an idea - might be good to allow a CSV to be drag & dropped directly onto the application shortcut to open and import...